backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
pydantic[email]
psycopg2
python-multipart
